Refactor modal and add quest components
