+++ components/dashboard/Welcome/WelcomeSection.tsx
@@ -4,13 +4,16 @@ import { GiJusticeStar } from "react-icons/gi";
 import { useRouter } from "next/navigation";
 import { BiLogOut } from "react-icons/bi";
 
-import DateDisplay from "@/components/dateDisplay";
+import DateDisplay from "@/components/DateDisplay";
 
 import { auth } from "@/firebase/firebase";
-import { UserInfo } from "@/models/user-schema";
+import { TUserInfo } from "@/types/users";
 
-
-export default function WelcomeSection({ currentUser }: { currentUser: UserInfo }) {
+export default function WelcomeSection({
+  currentUser,
+}: {
+  currentUser: TUserInfo;
+}) {
   const router = useRouter();
 
   const handleLogout = async () => {
@@ -26,7 +29,10 @@ export default function WelcomeSection({ currentUser }: { currentUser: UserInfo 
         <GiJusticeStar className="text-red-900 text-3xl mr-2 border-2 border-red-900 bg-white" />
         <h1 className="flex justify-between items-center text-xl text-red-900 w-full pl-2 border-2 border-red-900 bg-white">
           <span>
-          Welcome again, <strong>{currentUser ? currentUser.currentUser?.displayName : 'Guest'}</strong>
+            Welcome again,{" "}
+            <strong>
+              {currentUser ? currentUser.currentUser?.displayName : "Guest"}
+            </strong>
           </span>
           <button
             className="mr-1 p-0.5 hover:bg-rose-200 transition-colors duration-600 rounded-full"
